docker: switch base image from Alpine to Ubuntu-slim
Alpine Linux is based on Musl libc, but when enabling Fluent Bit
with Jemalloc in Musl, there are some issues when handling memory.

Since future features of Fluent Bit will use Jemalloc built-in metrics
of memory usage, so we are switching to Ubuntu-slim now to avoid
problems.

Ubuntu-slim in terms of size is greater than Alpine. This could change
again in the future.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>